card-generator/requirements.txt
Pillow==10.4.0
Flask==3.0.3
Flask-CORS==4.0.1
pytest==8.3.3
